+++ packages/expo-camera/src/Camera.tsx
@@ -109,6 +109,7 @@ export default class Camera extends React.Component<CameraProps> {
     Type: CameraManager.Type,
     FlashMode: CameraManager.FlashMode,
     AutoFocus: CameraManager.AutoFocus,
+    BarcodeType: CameraManager.BarcodeType,
     WhiteBalance: CameraManager.WhiteBalance,
     VideoQuality: CameraManager.VideoQuality,
     VideoStabilization: CameraManager.VideoStabilization || {},